+++ poi-scratchpad/src/main/java/org/apache/poi/hslf/dev/UserEditAndPersistListing.java
@@ -122,6 +122,6 @@ public final class UserEditAndPersistListing {
         long type = LittleEndian.getUShort(fileContents, pos+2);
         long rlen = LittleEndian.getUInt(fileContents, pos+4);
 
-        return Record.createRecordForType(type,fileContents,pos,(int)rlen+8);
+        return Record.createRecordForType(type,fileContents,pos,Math.toIntExact(rlen+8));
     }
 }
